refactor: extract orchestrator artifact validation seam
