+++ forgeflow_runtime/errors.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+
+class RuntimeViolation(Exception):
+    """Raised when a requested stage transition violates runtime policy."""
+
